Revert "refactor: transform SystemMessageLive into a component -> avoids having to handle :not_mounted_at_router case"
This reverts commit b29a30bb88b431c5c52823d005bed928b6e5097c.
