+++ lib/oli_web/components/layouts/instructor_dashboard.html.heex
@@ -1,5 +1,3 @@
-<.system_messages messages={assigns[:system_messages] || []} />
-
 <div class="flex-1 flex flex-col min-h-screen">
   <Components.Delivery.InstructorDashboard.header
     ctx={@ctx}
