+++ lib/oli_web/components/layouts/student_delivery.html.heex
@@ -1,5 +1,3 @@
-<.system_messages messages={assigns[:system_messages] || []} />
-
 <div class="h-screen flex flex-col overscroll-none">
   <Components.Delivery.Layouts.header
     ctx={@ctx}
