+++ lib/oli_web/components/layouts/student_delivery_lesson.html.heex
@@ -1,5 +1,3 @@
-<.system_messages messages={assigns[:system_messages] || []} />
-
 <div id="eventIntercept" phx-hook="LoadSurveyScripts">
   <div :if={Phoenix.LiveView.connected?(@socket) and assigns[:scripts_loaded]}>
     <script>
