+++ lib/oli_web/components/layouts/workspace.html.heex
@@ -1,5 +1,3 @@
-<.system_messages messages={assigns[:system_messages] || []} />
-
 <main role="main" class="relative flex flex-col">
   <div
     :if={assigns[:header_enabled?] && !@disable_sidebar? && assigns[:active_workspace]}
