+++ lib/oli_web/templates/layout/live.html.heex
@@ -49,6 +49,4 @@
 >
 </script>
 
-<.system_messages messages={assigns[:system_messages] || []} />
-
 <%= @inner_content %>
