+++ lib/oli_web/templates/static_page/index.html.eex
@@ -1,5 +1,3 @@
-<%= OliWeb.Components.Common.system_messages(%{messages: assigns[:system_messages] || []}) %>
-
 <div class="relative h-[calc(100vh-112px)] flex justify-center items-center">
     <div class="absolute h-[calc(100vh-112px)] w-full top-0 left-0">
       <%= OliWeb.Backgrounds.student_sign_in(%{}) %>
